# swagger.yaml
swagger: "2.0"
info:
  description: "Hotel CRUD based on MVC using node.js and express. Basic implementation of user, hotel, room, room services and reservation."
  version: "1.0.0"
  title: "Hotel CRUD"
host: "127.0.0.1:3000/"
...
schemes:
- "http"
paths:
...
  /user/fetchReservations:
    get:
      tags:
      - "user"
      summary: "Find User Reservations"
      description: ""
      operationId: "fetchUserReservations"
      produces:
      - "application/json"
      parameters:
        - name: token
          in: header
          type: string
          required: true
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/FetchReservationsResponse"
        "400":
          description: "Invalid data"
...
definitions:
  CreateReservationResponse:
    type: object
    properties:
        id:
          type: integer
          example: 1
        starting_date:
          type: string
          example: "2021-12-01T00:00:00.000Z"
        ending_date:
          type: string
          example: "2021-12-01T00:00:00.000Z"
        customer_id:
          type: integer
          example: 3
        createdAt:
          type: integer
          example: "2021-12-01T00:00:00.000Z"
        updatedAt:
          type: integer
          example: "2021-12-01T00:00:00.000Z"
        room_id:
          type: integer
          example: 2
...
  FetchReservationsResponse:
    type: "array"
    items:
      $ref: '#/definitions/CreateReservationResponse'
...